+++ Package.swift
@@ -1,0 +1,24 @@
+// swift-tools-version: 5.10
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "AntifraudSDK",
+    platforms: [.iOS(.v15)],
+    products: [
+        .library(
+            name: "AntifraudSDK",
+            targets: ["AntifraudSDK"]),
+    ],
+    dependencies: [
+    ],
+    targets: [
+        .binaryTarget(
+            name: "AntifraudSDK",
+            url: "https://github.com/bespot/antifraud-sdk-ios-release/releases/download/1.1.1/AntifraudSDK.xcframework.zip",
+            checksum: "b322ff909087f23648f437f3b1b1a827ed51a7b9915acd0fa0ff34dbff3f2a34"
+        )
+    ]
+)
+